Crate core_eth_relay_runtime

Modules

api

Structs

BlockHashCount

ChainForkVersions

CoreEthRelayDomainId

DummyTimeProvider

Dummy time provider that always returns zero.

ExistentialDeposit

GenesisConfig

MaxExtraDataSize

MaxFeeRecipientSize

MaxFinalizedHeaderSlotArray

MaxLocks

MaxLogsBloomSize

MaxProofBranchSize

MaxPublicKeySize

MaxReserves

MaxSignatureSize

MaxSlotsPerHistoricalRoot

MaxSyncCommitteeSize

MaximumRelayers

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

RelayConfirmationDepth

RelayerDeposit

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeOrigin

The runtime origin type representing the origin of a call.

SessionKeys

StateRootsBound

TransactionByteFee

TransporterEndpointId

Version

WeakSubjectivityPeriodSeconds

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

RuntimeEvent

Constants

EXISTENTIAL_DEPOSIT

The existential deposit. Same with the one on primary chain.

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

EthereumBeaconClient

EthereumBeaconClientConfig

Executive

Executive: handles dispatch to the various modules.

ExecutivePallet

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Index

Index of a transaction in the chain.

Messenger

MessengerConfig

RelayerId

The type we use to represent relayer id. This is same as account Id.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

Sudo

SudoConfig

System

SystemConfig

TransactionPayment

TransactionPaymentConfig

Transporter

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.